+++ www/packages/docs-utils/src/global-config.ts
@@ -4,8 +4,8 @@ import { DocsConfig } from "types"
 
 export const globalConfig: Pick<DocsConfig, "version"> = {
   "version": {
-    "number": "2.19",
-    "releaseUrl": "https://github.com/medusajs/medusa/releases/tag/v2.19.0",
-    "releaseDate": "2026-08-13T13:09:55Z"
+    "number": "2.21",
+    "releaseUrl": "https://github.com/medusajs/medusa/releases/tag/v2.21.0",
+    "releaseDate": "2026-09-11T10:13:57Z"
   }
 }